+++ homeassistant/components/sensor/tahoma.py
@@ -10,8 +10,7 @@ import logging
 from datetime import timedelta
 from homeassistant.helpers.entity import Entity
 from homeassistant.components.sensor import ENTITY_ID_FORMAT
-from homeassistant.components.tahoma import (
-    TAHOMA_DEVICES, TahomaDevice)
+from homeassistant.components.tahoma import (TAHOMA_DEVICES, TahomaDevice)
 
 DEPENDENCIES = ['tahoma']
 
@@ -22,7 +21,6 @@ SCAN_INTERVAL = timedelta(seconds=10)
 
 def setup_platform(hass, config, add_devices, discovery_info=None):
     """Set up the Tahoma controller devices."""
-    _LOGGER.error(TAHOMA_DEVICES['sensor'])
     add_devices(
         TahomaSensor(device, TAHOMA_DEVICES['api'])
         for device in TAHOMA_DEVICES['sensor'])
